Refactor X52ProManager
